GitLab.NET.Tests/RequestModels/GetDeployKeyRequestTests.cs
using GitLab.NET.RequestModels;
using GitLab.NET.Tests.TestHelpers;
using RestSharp;
using Xunit;

namespace GitLab.NET.Tests.RequestModels
{
    public class GetDeployKeyRequestTests
    {
        [Fact]
        public void GetRequest_AddsKeyIdParameter()
        {
            const uint expected = 5;
            var expectedParameter = new Parameter
            {
                Name = "keyId",
                Value = expected,
                Type = ParameterType.UrlSegment
            };
            var sut = new GetDeployKeyRequest(0, expected);

            var result = sut.GetRequest();

            Assert.Contains(expectedParameter, result.Parameters, new ParameterEqualityComparer());
        }

        [Fact]
        public void GetRequest_AddsProjectIdParameter()
        {
            const uint expected = 5;
            var expectedParameter = new Parameter
            {
                Name = "projectId",
                Value = expected,
                Type = ParameterType.UrlSegment
            };
            var sut = new GetDeployKeyRequest(expected, 0);

            var result = sut.GetRequest();

            Assert.Contains(expectedParameter, result.Parameters, new ParameterEqualityComparer());
        }

        [Fact]
        public void GetRequest_SetsCorrectResource()
        {
            var sut = new GetDeployKeyRequest(0, 0);

            var result = sut.GetRequest();

            Assert.Equal(GetDeployKeyRequest.Resource, result.Resource);
        }

        [Fact]
        public void GetRequest_SetsMethodToGet()
        {
            var sut = new GetDeployKeyRequest(0, 0);

            var result = sut.GetRequest();

            Assert.Equal(Method.GET, result.Method);
        }
    }
}